refactor: 크루 관리 입력 form 분리 및 폴더 구조 변경
